feat: move analysis functions to package
