+++ analysis/__init__.py
@@ -1,0 +1,6 @@
+from .portfolio_analysis import analyze_portfolio_improved, analyze_portfolio_async
+
+__all__ = [
+    "analyze_portfolio_improved",
+    "analyze_portfolio_async",
+]
