+++ reports/__init__.py
@@ -1,0 +1,17 @@
+from .reporting import (
+    print_analysis_results,
+    save_recommendations_to_db,
+    generate_analysis_report,
+    save_full_report,
+    generate_html_report,
+    save_html_report,
+)
+
+__all__ = [
+    "print_analysis_results",
+    "save_recommendations_to_db",
+    "generate_analysis_report",
+    "save_full_report",
+    "generate_html_report",
+    "save_html_report",
+]
